+++ src/Wpf.Ui.Violeta.Gallery/MainWindow.xaml.cs
@@ -173,6 +173,8 @@ public partial class MainWindow : ShellWindow
         ["navigation/tab-control"] = static () => new TabControlPage(),
         ["navigation/tab-view"] = static () => new TabViewPage(),
         ["navigation/tab-strip"] = static () => new TabStripPage(),
+        ["navigation/segmented"] = static () => new SegmentedPage(),
+        ["basic-input/segmented"] = static () => new SegmentedPage(),
         ["navigation/pivot"] = static () => new PivotPage(),
         ["navigation/anchor"] = static () => new AnchorPage(),